requirements.txt
absl-py==2.3.1
altair==6.0.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
attrs==24.3.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
cachetools==6.2.6
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.3
curl_cffi==0.13.0
cycler==0.12.0
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.24.3
fire==0.7.1
Flask==3.1.3
flask-cors==6.0.2
flatbuffers==25.12.19
fonttools==4.43.0
fqdn==1.5.1
frozendict==2.4.7
gast==0.6.0
gdown==5.2.1
gitdb==4.0.12
GitPython==3.1.46
google-pasta==0.2.0
grpcio==1.74.0
gunicorn==25.1.0
h11==0.14.0
h5py==3.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ImageHash==4.3.2
ImageIO==2.37.2
imbalanced-learn==0.13.0
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.11.3
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
lightdsa==0.0.3
lightecc==0.0.4
lightphe==0.0.20
lz4==4.4.5
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.8
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.9
mistune==3.1.0
ml_dtypes==0.5.3
mpmath==1.3.0
mtcnn==1.0.0
multitasking==0.0.12
namex==0.1.0
narwhals==2.17.0
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.1
notebook==7.3.2
notebook_shim==0.2.4
numpy==2.4.2
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
opt_einsum==3.4.0
optree==0.17.0
overrides==7.7.0
packaging==23.2
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.4
peewee==4.0.0
pillow==12.1.1
platformdirs==4.3.6
prometheus_client==0.21.1
prompt_toolkit==3.0.48
protobuf==3.20.3
psutil==6.1.1
pure_eval==0.2.3
pyarrow==23.0.1
pycparser==2.22
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.2.1
python-json-logger==3.2.1
pytz==2023.3.post1
PyWavelets==1.9.0
pywin32==308
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
retina-face==0.0.17
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.1.0
rpds-py==0.22.3
scikit-image==0.26.0
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
Send2Trash==1.8.3
six==1.16.0
sklearn-compat==0.1.3
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.5
soupsieve==2.6
split-folders==0.5.1
stack-data==0.6.3
streamlit==1.54.0
sympy==1.14.0
tenacity==9.1.4
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
terminado==0.18.1
textblob==0.19.0
threadpoolctl==3.5.0
tifffile==2026.2.24
tinycss2==1.4.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2023.3
uri-template==1.3.0
urllib3==2.3.0
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==16.0
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wordcloud==1.9.4
wrapt==1.17.3
yfinance==1.2.0
